requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
google-generativeai==0.3.2
whisperx==3.1.1
mediapipe==0.10.9
opencv-python==4.9.0.80
requests==2.31.0
yt-dlp==2024.1.1
jsonschema==4.20.0
